+++ .gitignore
@@ -50,5 +50,3 @@ notes.otl
 .nfs*
 
 node_modules
-
-credentials.conf
